- In a double monitor setup the Unity top panel in the second screen is cut at the right (LP:750481) - Unable to load icon text-x-preview at size 48 in a loop (LP: 794556)
(bzr r1184.3.16)

+++ src/PanelTray.h
@@ -50,7 +50,7 @@ public:
   void OnEntryRemoved (IndicatorObjectEntryProxy *proxy);
 
 public:
-  int        _n_children;
+  guint8     _n_children;
   char     **_whitelist;
 protected:
   const gchar * GetName ();
